requirements.txt
streamlit
langchain-huggingface
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
PyPDF2
python-docx